% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirf-package.R
\docType{package}
\name{dirf-package}
\alias{dirf}
\alias{dirf-package}
\title{dirf: Distances with Random Forest}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Predictions for mean and quantiles based on DiNo (Distance with Nodes) and RanBu (Random Bushes) methods.
}
\author{
\strong{Maintainer}: Tiago Mendonça \email{tiago.mendonca.s@gmail.com} (\href{https://orcid.org/0000-0002-4223-9162}{ORCID})

Authors:
\itemize{
  \item Rafael Izbicki \email{rizbicki@gmail.com}
}

}
\keyword{internal}
